+++ Sources/Sentry/SentryNSURLRequest.m
@@ -65,7 +65,6 @@ SentryNSURLRequest ()
     return self;
 }
 
-// TODO: Get refactored out to be a single init method
 - (_Nullable instancetype)initEnvelopeRequestWithDsn:(SentryDsn *)dsn
                                              andData:(NSData *)data
                                     didFailWithError:(NSError *_Nullable *_Nullable)error
@@ -103,9 +102,6 @@ newAuthHeader(NSURL *url)
         appendFormat:@"%@,",
         newHeaderPart(@"sentry_client",
             [NSString stringWithFormat:@"%@/%@", SentryMeta.sdkName, SentryMeta.versionString])];
-    [string
-        appendFormat:@"%@,",
-        newHeaderPart(@"sentry_timestamp", @((NSInteger)[[NSDate date] timeIntervalSince1970]))];
     [string appendFormat:@"%@", newHeaderPart(@"sentry_key", url.user)];
     if (nil != url.password) {
         [string appendFormat:@",%@", newHeaderPart(@"sentry_secret", url.password)];