+++ Tests/SentryTests/Helper/SentrySerializationTests.swift
@@ -22,6 +22,8 @@ class SentrySerializationTests: XCTestCase {
 
         let item = SentryEnvelopeItem(event: event)
         let envelope = SentryEnvelope(id: event.eventId, singleItem: item)
+        envelope.header.sentAt = Date(timeIntervalSince1970: 9_001)
+        
         // Sanity check
         XCTAssertEqual(event.eventId, envelope.header.eventId)
         XCTAssertEqual(1, envelope.items.count)
@@ -35,6 +37,7 @@ class SentrySerializationTests: XCTestCase {
             XCTAssertEqual(envelope.items[0].header.length, deserializedEnvelope.items[0].header.length)
             XCTAssertEqual(envelope.items[0].data, deserializedEnvelope.items[0].data)
             XCTAssertNil(deserializedEnvelope.header.traceContext)
+            XCTAssertEqual(Date(timeIntervalSince1970: 9_001), deserializedEnvelope.header.sentAt)
         }
     }