+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## Unreleased
+## 8.7.1
+
+### Features
+
+- Add `sent_at` to envelope header (#2859)
+
+### Fixes
+
+- Fix import of `User` & `Breadcrumb` (#3017)
+
+## 8.7.0
 
 ### Fixed